Move `runInternalSkikoComposeUiTest` to `v2.runInternalSkikoComposeUiTest`.
