+++ compose/ui/ui-test/src/skikoMain/kotlin/androidx/compose/ui/test/ComposeUiTest.skiko.kt
@@ -126,34 +126,6 @@ fun runSkikoComposeUiTest(
     ).runTest(block)
 }
 
-@InternalTestApi
-@OptIn(InternalComposeUiApi::class, ExperimentalTestApi::class, ExperimentalCoroutinesApi::class)
-fun runInternalSkikoComposeUiTest(
-    width: Int = 1024,
-    height: Int = 768,
-    density: Density = Density(1f),
-    effectContext: CoroutineContext = EmptyCoroutineContext,
-    runTestContext: CoroutineContext = EmptyCoroutineContext,
-    testTimeout: Duration = Duration.INFINITE,
-    semanticsOwnerListener: PlatformContext.SemanticsOwnerListener? = null,
-    windowInsets: PlatformWindowInsets? = null,
-    block: suspend SkikoComposeUiTest.() -> Unit,
-): TestResult {
-    return runTest {
-        SkikoComposeUiTest(
-            width = width,
-            height = height,
-            effectContext = effectContext,
-            runTestContext = runTestContext,
-            testTimeout = testTimeout,
-            density = density,
-            semanticsOwnerListener = semanticsOwnerListener,
-            windowInsets = windowInsets,
-            useStandardTestDispatcherForComposition = true,
-        ).runTest(block)
-    }
-}
-
 /**
  * How often to check idling resources.
  * Empirically checked that Android (espresso, really) tests approximately at this rate.
